Fixed basemap floodings in coastal areas

+++ OsmAndMapCreator/src/net/osmand/data/preparation/BasemapProcessor.java
@@ -276,7 +276,7 @@ public class BasemapProcessor {
             int x1 = x << (TILE_ZOOMLEVEL - zoom);
             int y1 = y << (TILE_ZOOMLEVEL - zoom);
             int max = 1 << (TILE_ZOOMLEVEL - zoom);
-            float c = 0;
+            int c = 0;
             for (int i = 0; i < max; i++) {
                 for (int j = 0; j < max; j++) {
 //                    if (landTileInfo.get((y1 + i) * 4096 + (x1 + j))) {
@@ -288,7 +288,6 @@ public class BasemapProcessor {
                 	if (!seaTileInfo.get((y1 + i) * 4096 + (x1 + j)) && !landTileInfo.get((y1 + i) * 4096 + (x1 + j))) {
                 		c --;
                 	}
-
                 }
             }
             float res = ((float)c) / ((float) max * (float) max);
